First test run successful
So the first test run was successful. I made some changes in
customInput.m file and saw the waveforms did get affected based on
that. Because of the run the program generated some other set of data.
Just committing those as well in case some thing goes wrong in the
future.